+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # Change Log
 
 ## 1.9.2
-- fix: URL generation handles parentheses and hyphens. Aligns with dev-portal doc viewer stub function
+- fix: URL param, camel case, and pascal case generation handles parentheses and hyphens. URL param stub aligns with dev-portal doc viewer stub function